+++ includes/class-testimonials-widget-widget-archives.php
@@ -20,21 +20,21 @@ with this program; if not, write to the Free Software Foundation, Inc.,
 
 require_once AIHR_DIR_INC . 'class-aihrus-widget.php';
 
-if ( class_exists( 'Testimonials_Widget_Widget_Categories' ) )
+if ( class_exists( 'Testimonials_Widget_Widget_Archives' ) )
 	return;
 
 
-class Testimonials_Widget_Widget_Categories extends Aihrus_Widget {
-	const ID = 'testimonials_widget_widget_categories';
+class Testimonials_Widget_Widget_Archives extends Aihrus_Widget {
+	const ID = 'testimonials_widget_widget_archives';
 
 	public static $title;
 
 
 	public function __construct() {
 		$classname   = __CLASS__;
-		$description = esc_html__( 'A list or dropdown of testimonials\' categories.', 'testimonials-widget' );
+		$description = esc_html__( 'A monthly archive of your site\'s testimonials.', 'testimonials-widget' );
 		$id_base     = self::ID;
-		self::$title = esc_html__( 'Testimonials Categories', 'testimonials-widget' );
+		self::$title = esc_html__( 'Testimonials Archives', 'testimonials-widget' );
 
 		parent::__construct( $classname, $description, $id_base, self::$title );
 	}
@@ -46,55 +46,48 @@ class Testimonials_Widget_Widget_Categories extends Aihrus_Widget {
 	public static function get_content( $instance, $widget_number ) {
 		$c = ! empty( $instance['count'] ) ? '1' : '0';
 		$d = ! empty( $instance['dropdown'] ) ? '1' : '0';
-		$h = ! empty( $instance['hierarchical'] ) ? '1' : '0';
-
-		$cat_args = array(
-			'orderby' => 'name',
-			'show_count' => $c,
-			'hierarchical' => $h,
-		);
-
-		$use_cpt_taxonomy = tw_get_option( 'use_cpt_taxonomy', false );
-		if ( $use_cpt_taxonomy ) {
-			$cat_args['taxonomy'] = Testimonials_Widget::$cpt_category;
-		}
+		
+		add_filter( 'getarchives_where', array( __CLASS__, 'getarchives_where' ), 10, 2 );
 
 		if ( $d ) {
-			$cat_args['show_option_none'] = esc_html__( 'Select Testimonials Category' );
-
-			wp_dropdown_categories( apply_filters( 'testimonials_widget_widget_categories_dropdown_args', $cat_args ) );
 			?>
-
-			<script type='text/javascript'>
-			/* <![CDATA[ */
-			var dropdown = document.getElementById("cat");
-			function onCatChange() {
-				if ( dropdown.options[dropdown.selectedIndex].value > 0 ) {
-					location.href = "<?php echo home_url(); ?>/?cat="+dropdown.options[dropdown.selectedIndex].value;
-				}
-			}
-			dropdown.onchange = onCatChange;
-			/* ]]> */
-			</script>
-
+			<select name="archive-dropdown" onchange='document.location.href=this.options[this.selectedIndex].value;'>
+				<option value=""><?php echo esc_attr( __( 'Select Month' ) ); ?></option>
+				<?php
+				wp_get_archives( apply_filters( 'testimonials_widget_widget_archives_dropdown_args', array(
+					'type'            => 'monthly',
+					'format'          => 'option',
+					'show_post_count' => $c
+				) ) );
+				?>
+			</select>
 			<?php
 		} else {
 			?>
 			<ul>
 			<?php
-				$cat_args['title_li'] = '';
-
-				wp_list_categories( apply_filters( 'testimonials_widget_widget_categories_args', $cat_args ) );
+				wp_get_archives( apply_filters( 'testimonials_widget_widget_archives_args', array(
+					'type'            => 'monthly',
+					'show_post_count' => $c
+				) ) );
 			?>
 			</ul>
 			<?php
 		}
+		
+		remove_filter( 'getarchives_where', array( __CLASS__, 'getarchives_where' ), 10, 2 );
 	}
 
 
 	/**
 	 * @SuppressWarnings(PHPMD.UnusedFormalParameter)
 	 */
+	public static function getarchives_where( $where, $args ) {
+		error_log( print_r( func_get_args(), true ) . ':' . __LINE__ . ':' . basename( __FILE__ ) );
+		return "WHERE post_type = '" . Testimonials_Widget::PT . "' AND post_status = 'publish'";
+	}
+
+
 	public static function form_parts( $instance = null, $number = null ) {
 		$form_parts = parent::form_parts( $instance, $number );
 
@@ -114,18 +107,11 @@ class Testimonials_Widget_Widget_Categories extends Aihrus_Widget {
 			'std' => 0,
 		);
 
-		$form_parts['hierarchical'] = array(
-			'title' => esc_html__( 'Show hierarchy', 'testimonials-widget' ),
-			'type' => 'checkbox',
-			'validate' => 'is_true',
-			'std' => 0,
-		);
-
 		foreach ( $form_parts as $id => $parts ) {
 			$form_parts[ $id ] = wp_parse_args( $parts, self::$default );
 		}
 
-		$form_parts = apply_filters( 'testimonials_widget_widget_categories_options', $form_parts );
+		$form_parts = apply_filters( 'testimonials_widget_widget_archives_options', $form_parts );
 
 		return $form_parts;
 	}