+++ includes/class-testimonials-widget-widget.php
@@ -27,6 +27,7 @@ if ( class_exists( 'Testimonials_Widget_Widget' ) )
 class Testimonials_Widget_Widget extends Aihrus_Widget {
 	const ID = 'testimonials_widget';
 
+
 	public function __construct() {
 		$classname   = __CLASS__;
 		$description = esc_html__( 'Display testimonials with multiple selection and display options', 'testimonials-widget' );