+++ includes/class-testimonials-widget.php
@@ -23,6 +23,7 @@ require_once AIHR_DIR_LIB . 'class-redrokk-metabox-class.php';
 require_once TW_DIR_INC . 'class-testimonials-widget-settings.php';
 require_once TW_DIR_INC . 'class-testimonials-widget-template-loader.php';
 require_once TW_DIR_INC . 'class-testimonials-widget-widget.php';
+require_once TW_DIR_INC . 'class-testimonials-widget-widget-archives.php';
 require_once TW_DIR_INC . 'class-testimonials-widget-widget-categories.php';
 require_once TW_DIR_INC . 'class-testimonials-widget-widget-tag-cloud.php';
 
@@ -1513,6 +1514,7 @@ EOF;
 
 	public static function widgets_init() {
 		register_widget( 'Testimonials_Widget_Widget' );
+		register_widget( 'Testimonials_Widget_Widget_Archives' );
 		register_widget( 'Testimonials_Widget_Widget_Categories' );
 		register_widget( 'Testimonials_Widget_Widget_Tag_Cloud' );
 	}